app/requirements-production.txt
django
djangorestframework
django-filter
djangorestframework-camel-case
django-health-check
python-dateutil
pytz
requests
whitenoise
mysqlclient
django-storages
boto3
django-redis-cache
uwsgi
awscli
pillow
